Separate broker endpoint to own files
