+++ components/kyma-environment-broker/internal/cluster_config/aws_provider.go
@@ -1,15 +1,13 @@
-package broker
+package cluster_config
 
 import (
 	"github.com/kyma-incubator/compass/components/kyma-environment-broker/internal"
 	"github.com/kyma-incubator/compass/components/provisioner/pkg/gqlschema"
 )
 
-type awsInputProvider struct{}
+type AwsInputProvider struct{}
 
-var _ hyperscalerInputProvider = &awsInputProvider{}
-
-func (p *awsInputProvider) Defaults() *gqlschema.ClusterConfigInput {
+func (p *AwsInputProvider) Defaults() *gqlschema.ClusterConfigInput {
 	return &gqlschema.ClusterConfigInput{
 		GardenerConfig: &gqlschema.GardenerConfigInput{
 			KubernetesVersion: "1.15.4",
@@ -36,6 +34,6 @@ func (p *awsInputProvider) Defaults() *gqlschema.ClusterConfigInput {
 	}
 }
 
-func (p *awsInputProvider) ApplyParameters(input *gqlschema.ClusterConfigInput, params internal.ProvisioningParametersDTO) {
+func (p *AwsInputProvider) ApplyParameters(input *gqlschema.ClusterConfigInput, params internal.ProvisioningParametersDTO) {
 	updateString(&input.GardenerConfig.ProviderSpecificConfig.AwsConfig.Zone, params.Zone)
 }
