+++ components/kyma-environment-broker/internal/cluster_config/common.go
@@ -1,0 +1,7 @@
+package cluster_config
+
+func updateString(toUpdate *string, value *string) {
+	if value != nil {
+		*toUpdate = *value
+	}
+}
