+++ components/kyma-environment-broker/internal/director/fake_client.go
@@ -1,11 +1,35 @@
 package director
 
-type fakeClient struct{}
+import (
+	"sync"
+)
 
-func NewFakeDirectorClient() *fakeClient {
-	return &fakeClient{}
+type fakeClient struct {
+	mu   sync.Mutex
+	URLs map[string]string
 }
 
+func NewFakeDirectorClient() *fakeClient {
+	return &fakeClient{
+		URLs: make(map[string]string, 0),
+	}
+}
+
+func (dc *fakeClient) SetConsoleURL(runtimeID, URL string) {
+	dc.mu.Lock()
+	defer dc.mu.Unlock()
+
+	dc.URLs[runtimeID] = URL
+}
+
+// Director Client methods
+
 func (dc *fakeClient) GetConsoleURL(accountID, runtimeID string) (string, error) {
+	for id, URL := range dc.URLs {
+		if runtimeID == id {
+			return URL, nil
+		}
+	}
+
 	return "", nil
 }
