+++ components/kyma-environment-broker/internal/ptr/ptr.go
@@ -3,3 +3,7 @@ package ptr
 func Bool(in bool) *bool {
 	return &in
 }
+
+func String(str string) *string {
+	return &str
+}
